refactor: move profile methods to separate class
